+++ CHANGES.md
@@ -1,3 +1,7 @@
+
+- [SeqKit v0.10.1](https://github.com/shenwei356/seqkit/releases/tag/v0.10.1)
+[![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/seqkit/v0.10.1/total.svg)](https://github.com/shenwei356/seqkit/releases/tag/v0.10.1)
+    - `seqkit fx2tab`: new option `-q/--avg-qual` for outputting average read quality
 - [SeqKit v0.10.0](https://github.com/shenwei356/seqkit/releases/tag/v0.10.0)
 [![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/seqkit/v0.10.0/total.svg)](https://github.com/shenwei356/seqkit/releases/tag/v0.10.0)
     - `seqkit`: report error when input is directory.